requirements.txt
findmy==0.7.5
paho-mqtt==2.1.0
flask==3.1.0